+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/right-panel/headless-treeview/index.ts
@@ -1,5 +1,5 @@
 export { TreeView } from "./TreeView";
 export { TreeItem } from "./TreeItem";
-export { createHeadlessDropHandler } from "./handleHeadlessDrop";
-export { createHeadlessCanDropHandler } from "./handleHeadlessCanDrop";
+export { createHeadlessDropHandler } from "./handlers/handleOnDrop";
+export { createHeadlessCanDropHandler } from "./handlers/handleCanDropAt";
 export type { TreeItemData, TreeItemInstance } from "./types";
